refactor(core): 将属性修改器数值类型从float改为int
- 修改StatModifier结构体中value字段的数据类型为int
- 移除了浮点数精度以提高性能和一致性

+++ include/core/Modifier.h
@@ -14,7 +14,7 @@
 struct StatModifier {
     StatType stat; // 修改哪个属性
     ModifyMode mode; // 修改方式
-    float value; // 修改数值
+    int value; // 修改数值
     EffectType effectType; // 正/负/中
 };
